refactor: move file upload constants to a separate module
- Create a new constants file for file upload types and defaults.
- Import DEFAULT_ALLOWED_TYPES in the FileUpload component.
- Clean up the FileUpload component by removing inline constants.
